chatbot-rag-poc/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Internal AI Chatbot API"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "psycopg2-binary",
    "alembic",
    "pydantic-settings",
    "pydantic[email]",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "httpx",
    "faiss-cpu",
    "llama-index-core",
    "llama-index-llms-ollama",
    "llama-index-embeddings-ollama",
    "pypdf",
    "python-docx",
    "python-multipart",
    "numpy",
    "aiofiles",
    "nest-asyncio",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
